+++ tests/Transformer/String/NormalizeWhitespace/NormalizeWhitespaceTransformerTest.php
@@ -1,0 +1,32 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ExtendsSoftware\ExaPHP\Transformer\String\NormalizeWhitespace;
+
+use ExtendsSoftware\ExaPHP\Transformer\String\NormalizeWhitespace\NormalizeWhitespaceTransformer;
+use PHPUnit\Framework\Attributes\CoversMethod;
+use PHPUnit\Framework\Attributes\Test;
+use PHPUnit\Framework\TestCase;
+
+#[CoversMethod(NormalizeWhitespaceTransformer::class, 'transform')]
+class NormalizeWhitespaceTransformerTest extends TestCase
+{
+    #[Test]
+    public function removesNonAsciiAlphanumericCharacters(): void
+    {
+        $transformer = new NormalizeWhitespaceTransformer();
+        $value = $transformer->transform("  A1Áβ٣\t\n\r\x00\x1F\x7F !@#-_ 　𐍈९");
+
+        $this->assertSame("A1Áβ٣ \x00\x1F\x7F !@#-_ 𐍈९", $value);
+    }
+
+    #[Test]
+    public function returnsValueWhenInputIsNotString(): void
+    {
+        $transformer = new NormalizeWhitespaceTransformer();
+        $value = $transformer->transform(123);
+
+        $this->assertSame(123, $value);
+    }
+}